+++ app/src/main/java/com/example/marvelherosapp/screens/MainScreen.kt
@@ -1,0 +1,95 @@
+package com.example.marvelherosapp.screens
+
+import android.content.Context
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.aspectRatio
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.lazy.rememberLazyListState
+import androidx.compose.material.Text
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.snapshotFlow
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.navigation.NavHostController
+
+import kotlinx.coroutines.launch
+
+import com.example.marvelherosapp.utils.getCharacterName
+import com.example.marvelherosapp.utils.getImagePainter
+
+
+@Composable
+fun MainScreen(context: Context, navController: NavHostController)
+{
+    val listState = rememberLazyListState()
+    val coroutineScope = rememberCoroutineScope()
+
+    LaunchedEffect(listState) {
+        snapshotFlow { listState.firstVisibleItemIndex }
+            .collect { index ->
+                coroutineScope.launch {
+                    listState.animateScrollToItem(index)
+                }
+            }
+    }
+
+    LazyRow(
+        state = listState,
+        modifier = Modifier.fillMaxSize()
+    )
+    {
+        items(3) { index ->
+            Box(
+                modifier = Modifier
+                    .height(300.dp)
+                    .width(300.dp)
+                    .padding(8.dp)
+                    .clickable
+                    {
+                        navController.navigate("detail/$index")
+                    }
+            ) {
+                Image(
+                    painter = getImagePainter(index),
+                    contentDescription = "Hero number $index",
+                    contentScale = ContentScale.Crop,
+                    modifier = Modifier
+                        .fillMaxSize()
+                        .aspectRatio(1f)
+                )
+
+                Box(
+                    modifier = Modifier
+                        .align(Alignment.BottomCenter)
+                        .background(Color.Black.copy(alpha = 0.7f))
+                        .fillMaxWidth()
+                        .padding(8.dp),
+                    contentAlignment = Alignment.Center
+                ) {
+                    Text(
+                        text = getCharacterName(context, index),
+                        color = Color.White,
+                        fontSize = 20.sp,
+                        fontWeight = FontWeight.Bold
+                    )
+                }
+            }
+        }
+
+    }
+}
