+++ app/src/main/java/com/example/marvelherosapp/utils/Utils.kt
@@ -2,7 +2,6 @@ package com.example.marvelherosapp.utils
 
 
 import android.content.Context
-import android.content.res.Resources
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.res.painterResource
